Added: Author mappings, methods, model Modified: replaced manual added mapping with autogenerated mappings

+++ src/main/java/org/linkstorage/controller/LinkController.java
@@ -24,7 +24,7 @@ public class LinkController {
 
 	@RequestMapping(method=RequestMethod.GET, value="/link/{id}")
 	public ModelAndView getLink(@PathVariable String id) {
-		Link link = linksRepository.getLink(Long.parseLong(id));
+		Link link = linksRepository.getLink(Integer.parseInt(id));
 	    ModelAndView modelAndView = new ModelAndView(XML_VIEW_NAME);
 	    modelAndView.addObject("link", link);
 	    return modelAndView;
