+++ src/main/java/org/linkstorage/repository/HibernateLinkBase.java
@@ -1,13 +1,12 @@
 package org.linkstorage.repository;
 
+import java.util.List;
+
 import org.linkstorage.model.Link;
 import org.hibernate.SessionFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
 
 @Repository
 @Transactional
@@ -18,11 +17,11 @@ public class HibernateLinkBase implements LinkBase {
 	@Autowired
 	public HibernateLinkBase(SessionFactory sessionFactory) {
 	    this.sessionFactory = sessionFactory;
-	  }
+	}
 
 	@Override
 	@Transactional
-	public Link getLink(Long id) {
+	public Link getLink(Integer id) {
 	    return (Link) sessionFactory.getCurrentSession().get(Link.class, id);
 	}
